changes to make dust work for bursts
Main changes are to ensure the dust mass does not go -ve when you have a very strong burst.
Other changes - put in isotop calculations to use implicit arrays rather than the inline loops.
Also turned on the option to remove extra whitespace in spyder, so most differences are whitespace. To much hassle to go through and identify each one.